refactor server.js to move logic into meals_controller and meal model
